refactor: move manage domain verification to admin-console, refs AC-1202
